fix refresh scripts

+++ git/refresh-og.sh
@@ -2,6 +2,11 @@
 SCRIPT_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
 MY_NAME=joequant
 cd $SCRIPT_DIR
+if [ "$1" == "--upload" ] ; then
+export UPLOAD=1
+else
+export UPLOAD=0
+fi
 
 pushd ../..
 pushd OG-PlatformNative
@@ -10,14 +15,14 @@ git fetch origin
 git stash
 git checkout master
 git rebase upstream/master
-git push --set-upstream origin master
+if [ $UPLOAD -eq 1 ] ; then git push --set-upstream origin master ; fi
 git checkout develop
 git rebase upstream/develop
-git push --set-upstream origin develop
+if [ $UPLOAD -eq 1 ] ; then git push --set-upstream origin develop ; fi
 git checkout bitquant
 git rebase origin/bitquant
 git rebase upstream/develop
-git push --force --set-upstream origin bitquant
+if [ $UPLOAD -eq 1 ] ; then git push --force --set-upstream origin bitquant ; fi
 popd
 
 pushd OG-Platform
@@ -26,12 +31,12 @@ git fetch origin
 git stash
 git checkout master
 git rebase upstream/master
-git push --set-upstream origin master
+if [ $UPLOAD -eq 1 ] ; then git push --set-upstream origin master ; fi
 git checkout develop
 git rebase upstream/develop
-git push --set-upstream origin develop
+if [ $UPLOAD -eq 1 ] ; then git push --set-upstream origin develop ; fi
 git checkout bitquant
 git rebase origin/bitquant
 git rebase upstream/develop
-git push --force --set-upstream origin bitquant
+if [ $UPLOAD -eq 1 ] ; then git push --force --set-upstream origin bitquant ; fi
 popd
